+++ tests/querygen-compile-run/build.rs
@@ -47,6 +47,11 @@ fn main() {
             r#"queries::AllFilms::fragment(FragmentContext::empty())"#,
         ),
         TestCase::query(
+            &starwars_schema,
+            "../../cynic-querygen/tests/queries/starwars/fragment-spreads.graphql",
+            r#"queries::AllFilms::fragment(FragmentContext::empty())"#,
+        ),
+        TestCase::query(
             &jobs_schema,
             "tests/queries/graphql.jobs/london-jobs.graphql",
             r#"queries::Jobs::fragment(FragmentContext::empty())"#,