+++ CHANGELOG.md
@@ -22,6 +22,7 @@ all APIs might be changed.
 
 - Cynic now supports recursive queries via the `#[cynic(recurse="N")]`
   attribute on fields that recurse.
+- The generator now understands query fragments & spreads.
 
 ### Bug Fixes